refactor: move dashboard to home and seperate components
