+++ app/api/todo/route.ts
@@ -11,7 +11,7 @@ export async function GET(request: NextRequest) {
   if (!query_date) {
     return NextResponse.json(
       { error: 'query_date parameter is required' },
-      { status: 400 },
+      { status: 400 }
     );
   }
 
@@ -28,8 +28,8 @@ export async function GET(request: NextRequest) {
       date.getUTCDate(),
       0,
       0,
-      0,
-    ),
+      0
+    )
   );
   const endOfDayUtc = new Date(
     Date.UTC(
@@ -38,8 +38,8 @@ export async function GET(request: NextRequest) {
       date.getUTCDate(),
       23,
       59,
-      59,
-    ),
+      59
+    )
   );
 
   try {
@@ -53,14 +53,14 @@ export async function GET(request: NextRequest) {
     });
 
     const tasksWithIdAsString = JSON.parse(
-      JSON.stringify(tasks, bigIntReplacer),
+      JSON.stringify(tasks, bigIntReplacer)
     );
 
     return NextResponse.json(tasksWithIdAsString, { status: 200 });
   } catch (error) {
     return NextResponse.json(
       { error: 'Failed to fetch tasks: ' + error.message },
-      { status: 500 },
+      { status: 500 }
     );
   }
 }
@@ -73,7 +73,7 @@ export async function PATCH(request: NextRequest) {
     if (!taskId || typeof completed !== 'boolean') {
       return NextResponse.json(
         { error: 'Invalid or missing taskId or completed status' },
-        { status: 400 },
+        { status: 400 }
       );
     }
 
@@ -112,14 +112,14 @@ export async function PATCH(request: NextRequest) {
     });
 
     const updatedTaskWithStringId = JSON.parse(
-      JSON.stringify(updatedTask, bigIntReplacer),
+      JSON.stringify(updatedTask, bigIntReplacer)
     );
 
     return NextResponse.json(updatedTaskWithStringId, { status: 200 });
   } catch (error) {
     return NextResponse.json(
       { error: 'Failed to update task: ' + (error as Error).message },
-      { status: 500 },
+      { status: 500 }
     );
   }
 }
